# openapi/poveventcam.yaml
openapi: 3.0.3
info:
  title: POV EventCamera API
  version: 0.1.0-draft
  description: |
    Draft OpenAPI contract for POV Cam guest and organizer APIs.
    This spec mirrors the current architecture docs and is intended as the
    starting point for typed client generation and server contract tests.
servers:
  - url: https://eventpovcamera.app
    description: Organizer web entry domain
  - url: https://guest.eventpovcamera.app
    description: Guest web entry domain
  - url: http://localhost:3000
    description: Local development
...
paths:
  /api/join:
    post:
      tags: [Guest]
      summary: Join an event and create a guest device session
      operationId: guestJoinEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRequest'
      responses:
        '200':
          description: Joined successfully
          headers:
            Set-Cookie:
              description: HttpOnly host-only cookie for guest.eventpovcamera.app containing device session token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/my-session:
    get:
      tags: [Guest]
      summary: Validate guest session and return session + event metadata
      operationId: guestGetMySession
      security:
        - GuestSessionCookie: []
      responses:
        '200':
          description: Session is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MySessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Guest]
      summary: Update guest display name
      operationId: guestPatchMySession
      security:
        - GuestSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMySessionRequest'
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MySessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/create-upload:
    post:
      tags: [Guest]
      summary: Reserve a media upload slot and return signed upload URLs
      operationId: guestCreateUpload
      security:
        - GuestSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadRequest'
      responses:
        '200':
          description: Upload slot reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/complete-upload:
    post:
      tags: [Guest]
      summary: Verify uploaded object and mark media as uploaded
      operationId: guestCompleteUpload
      security:
        - GuestSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '200':
          description: Upload completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/my-uploads:
    get:
      tags: [Guest]
      summary: List uploads that belong to the current guest session
      operationId: guestGetMyUploads
      security:
        - GuestSessionCookie: []
      responses:
        '200':
          description: Upload list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyUploadsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/organizer/events/{id}/guests:
    get:
      tags: [OrganizerGuests]
      summary: List event guest sessions
      operationId: organizerGetGuests
      security:
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Guest session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    GuestSessionCookie:
      type: apiKey
      in: cookie
      name: device_session_token
    OrganizerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UPLOAD_LIMIT_REACHED
            message:
              type: string
              example: You have reached your upload limit.
            request_id:
              type: string
              nullable: true
              example: req_12345
            details:
              type: object
              additionalProperties: true
    CompressionMode:
      type: string
      enum: [compressed, raw]
    EventStatus:
      type: string
      enum: [draft, active, closed, archived, purged]
...
    GuestSession:
      type: object
      required:
        - session_id
        - event_id
        - is_active
        - created_at
        - last_active_at
      properties:
        session_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
          maxLength: 64
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_active_at:
          type: string
          format: date-time
    GuestEventMeta:
      type: object
      required:
        - id
        - slug
        - name
        - status
        - max_uploads_per_guest
        - max_guests
        - compression_mode
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/EventStatus'
        max_uploads_per_guest:
          type: integer
        max_guests:
          type: integer
        compression_mode:
          $ref: '#/components/schemas/CompressionMode'
        expires_at:
          type: string
          format: date-time
    JoinRequest:
      type: object
      required: [event_slug]
      properties:
        event_slug:
          type: string
        pin:
          type: string
          pattern: '^\\d{4}$'
          nullable: true
        display_name:
          type: string
          maxLength: 64
          nullable: true
    JoinResponse:
      type: object
      required: [session, event, upload_count, max_uploads]
      properties:
        session:
          $ref: '#/components/schemas/GuestSession'
        event:
          $ref: '#/components/schemas/GuestEventMeta'
        upload_count:
          type: integer
          minimum: 0
        max_uploads:
          type: integer
          minimum: 1
    MySessionResponse:
      type: object
      required: [session, event, upload_count, max_uploads]
      properties:
        session:
          $ref: '#/components/schemas/GuestSession'
        event:
          $ref: '#/components/schemas/GuestEventMeta'
        upload_count:
          type: integer
          minimum: 0
        max_uploads:
          type: integer
          minimum: 1
    PatchMySessionRequest:
      type: object
      required: [display_name]
      properties:
        display_name:
          type: string
          nullable: true
          maxLength: 64
    CreateUploadRequest:
      type: object
      required: [file_type, file_size]
      properties:
        file_type:
          type: string
          description: MIME type declared by client
          example: image/jpeg
        file_size:
          type: integer
          minimum: 1
          description: File size in bytes declared by client
        tags:
          type: array
          description: Optional per-file labels provided by guest
          items:
            type: string
            maxLength: 32
          maxItems: 8
    CreateUploadResponse:
      type: object
      required:
        - media_id
        - upload_url
        - remaining_uploads
        - compression_mode
        - max_file_size
      properties:
        media_id:
          type: string
          format: uuid
        upload_url:
          type: string
          format: uri
        thumb_upload_url:
          type: string
          format: uri
          nullable: true
        remaining_uploads:
          type: integer
          minimum: 0
        compression_mode:
          $ref: '#/components/schemas/CompressionMode'
        max_file_size:
          type: integer
          minimum: 1
          description: Max allowed size in bytes for the reserved upload mode
    CompleteUploadRequest:
      type: object
      required: [media_id]
      properties:
        media_id:
          type: string
          format: uuid
    CompleteUploadResponse:
      type: object
      required: [success, media_id, status]
      properties:
        success:
          type: boolean
          example: true
        media_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [uploaded]
    MyUploadItem:
      type: object
      required: [media_id, thumb_url, status, uploaded_at]
      properties:
        media_id:
          type: string
          format: uuid
        thumb_url:
          type: string
          format: uri
        status:
          type: string
          enum: [uploaded, pending, failed, expired, hidden]
        uploaded_at:
          type: string
          format: date-time
          nullable: true
        uploader_name:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
    MyUploadsResponse:
      type: object
      required: [uploads, total_uploaded, max_uploads]
      properties:
        uploads:
          type: array
          items:
            $ref: '#/components/schemas/MyUploadItem'
        total_uploaded:
          type: integer
          minimum: 0
        max_uploads:
          type: integer
          minimum: 1
...
    GuestListItem:
      type: object
      required:
        - session_id
        - upload_count
        - is_active
        - created_at
        - last_active_at
      properties:
        session_id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        upload_count:
          type: integer
          minimum: 0
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_active_at:
          type: string
          format: date-time
    GuestsResponse:
      type: object
      required: [guests, total_guests, max_guests]
      properties:
        guests:
          type: array
          items:
            $ref: '#/components/schemas/GuestListItem'
        total_guests:
          type: integer
          minimum: 0
        max_guests:
          type: integer
          minimum: 1
...